Separate Plantmaster components into files
