+++ tgui/packages/tgui/interfaces/Plantmaster/tabs/index.ts
@@ -1,0 +1,11 @@
+/**
+ * @file
+ * @copyright 2022
+ * @author Original Amylizzle (https://github.com/amylizzle)
+ * @author Changes Mordent (https://github.com/mordent-goonstation)
+ * @license MIT
+ */
+
+export { ExtractablesView } from './ExtractablesView';
+export { OverviewView } from './OverviewView';
+export { SeedsView } from './SeedsView';
